IGNITE-19650 Fixes review issues.
